add some features and fix on post, and adding LOADER on CMS

+++ admin/includes/view_all_posts.php
@@ -14,9 +14,34 @@
                 confirmQuery($update_to_draft_status);
             break;
             case 'delete':
-                $query = "DELETE from posts WHERE post_id = '{$postValueId}' ";
+                $query = "DELETE FROM posts WHERE post_id = '{$postValueId}' ";
                 $update_to_delete_status = mysqli_query($connection, $query);
                 confirmQuery($update_to_delete_status);
+            break;
+            case 'clone':
+                $query = "SELECT * FROM posts WHERE post_id = '{$postValueId}' ";
+                $select_post_query = mysqli_query($connection, $query);
+
+                while($row = mysqli_fetch_array($select_post_query)){
+                    $post_title= $row['post_title'];
+                    $post_category_id= $row['post_category_id'];
+                    $post_date= $row['post_date'];
+                    $post_author= $row['post_author'];
+                    $post_status= $row['post_status'];
+                    $post_image= $row['post_image'];
+                    $post_tags= $row['post_tags'];
+                    $post_content = $row['post_content'];
+                }
+                $query = "INSERT INTO posts (post_category_id, post_title, post_author, post_date, post_image, post_content, post_tags,
+                                 post_status) ";
+                $query .= "VALUES ({$post_category_id}, '{$post_title}', '{$post_author}', now(), '{$post_image}', '{$post_content}','{$post_tags}',  '{$post_status}')";
+                $copy_query = mysqli_query($connection, $query);
+
+                if(!$copy_query){
+                    die("QUERY FAILED" . mysqli_error($connection));
+                }
+
+                confirmQuery($select_post_query);
             break;
             }
         }
@@ -33,6 +58,7 @@
                 <option value="published">Publish</option>
                 <option value="draft">Draft</option>
                 <option value="delete">Delete</option>
+                <option value="clone">Clone</option>
             </select>
         </div>
 
@@ -60,7 +86,7 @@
         </thead>
         <tbody>
             <?php 
-                $query = "SELECT * FROM posts";
+                $query = "SELECT * FROM posts ORDER BY post_id DESC";
                 $select_posts= mysqli_query($connection, $query);
 
                 while($row = mysqli_fetch_assoc($select_posts)){
@@ -97,7 +123,7 @@
                     echo "<td>{$post_date}</td>";
                     echo "<td><a href='../post.php?p_id={$post_id}'>View</a></td>";
                     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                    echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                    echo "<td><a onClick=\"javascript:return confirm('Are you sure want to delete?')\" href='posts.php?delete={$post_id}'>Delete</a></td>";
                     echo "</tr>";
                 }
             ?>
